# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Education Planner API",
    "description": "\n    ## Динамический планировщик образовательной траектории\n    \n    API для студентов УрФУ, помогающее:\n    - 📚 Выбирать специальность и курсы\n    - 📊 Отслеживать развитие навыков\n    - 🎯 Получать персональные рекомендации\n    - 🔄 Моделировать смену специальности\n    \n    ### Основные возможности:\n    \n    - **Аутентификация**: JWT токены (access + refresh)\n    - **Специальности**: Просмотр roadmap и матрицы навыков\n    - **Курсы**: Запись, прохождение, отслеживание прогресса\n    - **Навыки**: Древовидная структура, gap-анализ\n    - **Рекомендации**: Умные подсказки по выбору курсов\n    - **What-If анализ**: Моделирование смены специальности\n    \n    ### Для начала работы:\n    1. Зарегистрируйтесь через `/api/auth/register`\n    2. Получите токен через `/api/auth/login`\n    3. Используйте токен в заголовке: `Authorization: Bearer \u003Ctoken\u003E`\n    ",
    "contact": {
      "name": "Education Planner Team",
      "email": "support@edu-planner.ru"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/students/me/profile": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get My Profile",
        "description": "Получить профиль текущего студента",
        "operationId": "get_my_profile_api_students_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "students"
        ],
        "summary": "Update My Profile",
        "description": "Обновить профиль студента",
        "operationId": "update_my_profile_api_students_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/students/me/specialization": {
      "put": {
        "tags": [
          "students"
        ],
        "summary": "Set Specialization",
        "description": "Выбрать или сменить специальность",
        "operationId": "set_specialization_api_students_me_specialization_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "specialization_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Specialization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/students/me/courses": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get My Courses",
        "description": "Получить курсы студента: завершенные, текущие и доступные",
        "operationId": "get_my_courses_api_students_me_courses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get My Courses Api Students Me Courses Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/students/me/progress": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get My Progress",
        "description": "Получить прогресс обучения студента",
        "operationId": "get_my_progress_api_students_me_progress_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get My Progress Api Students Me Progress Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/students/me/resume": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get My Resume",
        "description": "Получить резюме студента на основе изученных курсов и навыков",
        "operationId": "get_my_resume_api_students_me_resume_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeEducation": {
        "properties": {
          "completed_courses": {
            "type": "integer",
            "title": "Completed Courses"
          },
          "total_credits": {
            "type": "integer",
            "title": "Total Credits"
          },
          "average_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Grade"
          }
        },
        "type": "object",
        "required": [
          "completed_courses",
          "total_credits"
        ],
        "title": "ResumeEducation",
        "description": "Образование для резюме"
      },
      "ResumeMinor": {
        "properties": {
          "minor_name": {
            "type": "string",
            "title": "Minor Name"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "selected, completed, archived"
          },
          "skills_gained": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Gained",
            "description": "Навыки из этого майнера"
          }
        },
        "type": "object",
        "required": [
          "minor_name",
          "status"
        ],
        "title": "ResumeMinor",
        "description": "Информация о майнере в резюме"
      },
      "ResumePersonalInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "specialization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "title": "ResumePersonalInfo",
        "description": "Личная информация для резюме"
      },
      "ResumeResponse": {
        "properties": {
          "personal": {
            "$ref": "#/components/schemas/ResumePersonalInfo"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/ResumeSkill"
            },
            "type": "array",
            "title": "Skills"
          },
          "minors": {
            "items": {
              "$ref": "#/components/schemas/ResumeMinor"
            },
            "type": "array",
            "title": "Minors",
            "default": []
          },
          "education": {
            "$ref": "#/components/schemas/ResumeEducation"
          },
          "achievements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Achievements"
          }
        },
        "type": "object",
        "required": [
          "personal",
          "skills",
          "education",
          "achievements"
        ],
        "title": "ResumeResponse",
        "description": "Полное резюме студента (основа + орбита)"
      },
      "ResumeSkill": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "type": "integer",
            "maximum": 3,
            "minimum": 0,
            "title": "Level",
            "description": "Уровень навыка (0-3)"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Источник: core (основная спец) или orbit (майнер)"
          },
          "acquired_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired From",
            "description": "Название майнера или спецы"
          }
        },
        "type": "object",
        "required": [
          "name",
          "level",
          "source"
        ],
        "title": "ResumeSkill",
        "description": "Навык в резюме с источником"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Full Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          },
          "minor_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Minor Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Схема ответа с данными пользователя"
      },
      "UserUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Схема для обновления профиля пользователя"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}